requirements.txt
python-telegram-bot==20.7
anthropic>=0.25.0
python-dotenv
